fix(api): merge rna siren files after rebase with file rename
